+++ web/js/037-map.js
@@ -216,6 +216,7 @@
 										}
 									}
 								}
+								self.numsMedia = self.media.imagesAndVideosCount();
 								resolve_cacheBasePromise();
 							},
 							function() {